pyproject.toml
[tool.poetry]
name = "rankwise"
version = "0.1.0"
description = "A tool that evaluates and ranks embedding models based on their retrieval performance using several metrics across various queries and document sets"
authors = ["Roberto Abdelkader Martínez Pérez <robertomartinezp@gmail.com>", "Pedro Ruiz Pareja <pedro.ruiz.tech@bbva.com>"]
license = "BSD"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
llama-index-core = "^0.11.10"
llama-index-llms-ollama = "^0.3.2"
llama-index-embeddings-ollama = "^0.3.1"
llama-index-embeddings-azure-openai = "^0.2.5"
llama-index-llms-azure-openai = "^0.2.1"

[tool.poetry.scripts]
rankwise = "rankwise.cli:main"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
